Refactor authentication and menu item services to use interfaces and implement service classes
