+++ restaurant-management-service-with-springboot/restaurant-management-system/src/main/java/com/restaurant/service/MenuItemService.java
@@ -1,37 +1,9 @@
 package com.restaurant.service;
 
 import com.restaurant.dto.MenuItemDTO;
-import com.restaurant.entity.MenuItem;
-import com.restaurant.repository.MenuItemRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.util.List;
+public interface MenuItemService {
+    void addMenuItem(MenuItemDTO request);
 
-@Service
-public class MenuItemService {
-
-    private final MenuItemRepository menuItemRepository;
-
-    public MenuItemService(MenuItemRepository menuItemRepository) {
-        this.menuItemRepository = menuItemRepository;
-    }
-
-    public void addMenuItem(MenuItemDTO request) {
-        MenuItem menuItem = MenuItem.builder()
-                .name(request.getName())
-                .price(request.getPrice())
-                .build();
-        menuItemRepository.save(menuItem);
-    }
-
-    public void deleteMenuItem(Long id) {
-        MenuItem menuItem = menuItemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Menu item not found"));
-        menuItemRepository.delete(menuItem);
-    }
-
-    public List<MenuItem> getAllMenuItems() {
-        return menuItemRepository.findAll();
-    }
+    void deleteMenuItem(Long id);
 }
