+++ restaurant-management-service-with-springboot/restaurant-management-system/src/main/java/com/restaurant/service/TableBookingService.java
@@ -3,67 +3,17 @@ package com.restaurant.service;
 import com.restaurant.dto.StatusDTO;
 import com.restaurant.dto.TableBookingDTO;
 import com.restaurant.entity.TableBooking;
-import com.restaurant.enumeration.BookingStatus;
-import com.restaurant.repository.TableBookingRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-public class TableBookingService {
-    private final TableBookingRepository tableBookingRepository;
+public interface TableBookingService {
+    List<TableBooking> getAllBookings();
 
-    public TableBookingService(TableBookingRepository tableBookingRepository) {
-        this.tableBookingRepository = tableBookingRepository;
-    }
+    TableBooking getBookingById(Long id);
 
-    public Long bookTable(TableBookingDTO request) {
-        TableBooking tableBooking = TableBooking.builder()
-                .customerName(request.getCustomerName())
-                .phoneNumber(request.getPhoneNumber())
-                .bookingDate(request.getBookingDate())
-                .NumberOfGuests(request.getNumberOfGuests())
-                .bookingStatus(BookingStatus.PENDING)
-                .build();
-        return tableBookingRepository.save(tableBooking).getId();
-    }
+    void updateBooking(Long id, TableBookingDTO request);
 
-    public List<TableBooking> getAllBookings() {
-        return tableBookingRepository.findAll();
-    }
+    void deleteBooking(Long id);
 
-    public TableBooking getBookingById(Long id) {
-        return tableBookingRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Booking not found with id: " + id));
-    }
-
-    public void updateBooking(Long id, TableBookingDTO request) {
-        TableBooking existingBooking = getBookingById(id);
-        if(request.getCustomerName()!=null) {
-            existingBooking.setCustomerName(request.getCustomerName());
-        }
-        if(request.getPhoneNumber()!=null) {
-            existingBooking.setPhoneNumber(request.getPhoneNumber());
-        }
-        if(request.getBookingDate()!=null) {
-            existingBooking.setBookingDate(request.getBookingDate());
-        }
-        if(request.getNumberOfGuests()!=null){
-        existingBooking.setNumberOfGuests(request.getNumberOfGuests());
-        }
-        tableBookingRepository.save(existingBooking);
-    }
-
-    public void deleteBooking(Long id) {
-        TableBooking existingBooking = getBookingById(id);
-        tableBookingRepository.delete(existingBooking);
-    }
-
-    public void updateBookingStatus(Long id, StatusDTO status) {
-        TableBooking existingBooking = getBookingById(id);
-        BookingStatus bookingStatus = BookingStatus.valueOf(status.getStatus().toUpperCase());
-        existingBooking.setBookingStatus(bookingStatus);
-        tableBookingRepository.save(existingBooking);
-    }
+    void updateBookingStatus(Long id, StatusDTO status);
 }
