+++ restaurant-management-service-with-springboot/restaurant-management-system/src/main/java/com/restaurant/service/impl/MenuItemServiceImpl.java
@@ -1,0 +1,39 @@
+package com.restaurant.service.impl;
+
+import com.restaurant.dto.MenuItemDTO;
+import com.restaurant.entity.MenuItem;
+import com.restaurant.repository.MenuItemRepository;
+import com.restaurant.service.MenuItemService;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class MenuItemServiceImpl implements MenuItemService {
+
+    private final MenuItemRepository menuItemRepository;
+
+    public MenuItemServiceImpl(MenuItemRepository menuItemRepository) {
+        this.menuItemRepository = menuItemRepository;
+    }
+
+    @Override
+    public void addMenuItem(MenuItemDTO request) {
+        MenuItem menuItem = MenuItem.builder()
+                .name(request.getName())
+                .price(request.getPrice())
+                .build();
+        menuItemRepository.save(menuItem);
+    }
+
+    @Override
+    public void deleteMenuItem(Long id) {
+        MenuItem menuItem = menuItemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Menu item not found"));
+        menuItemRepository.delete(menuItem);
+    }
+
+    public List<MenuItem> getAllMenuItems() {
+        return menuItemRepository.findAll();
+    }
+}
