+++ restaurant-management-service-with-springboot/restaurant-management-system/src/test/java/com/restaurant/service/AuthServiceImplTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
-public class AuthServiceTest {
+public class AuthServiceImplTest {
     @BeforeEach
     public void setUp() {
         // Initialize any required components or mocks here
